MAke more code just use the single shipment of the order even if the DB still supports multiple shipments per order
